consulta_final

+++ FinalFour/Models/Pick.cs
@@ -3,11 +3,13 @@ namespace FinalFour.Models
 {
     public class Pick
     {
-        public string? UID { get; set; }  
-        public string?Id { get; set; }
-        public string?Apuesta { get; set; }
-        public int?Momio { get; set; }
-        public int?Monto { get; set; }
-        public int?Resultado { get; set; }
+        public string? UID { get; set; }
+        public string? Id { get; set; }
+        public string? Apuesta { get; set; }
+        public float? Momio { get; set; }
+        public float? Monto { get; set; }
+        public int? Resultado { get; set; }
+
+        
     }
 }
